% Generated by roxygen2 (4.0.2): do not edit by hand
\name{centers_ca}
\alias{centers_ca}
\title{Compute Coordinates for Cluster Centers}
\usage{
centers_ca(x, clusters, freq)
}
\arguments{
\item{x}{A numeric matrix.}

\item{clusters}{A clustering of the row levels of \code{x}: either a list or the output of \code{kmeans}.}

\item{freq}{An optional vector of frequency counts for the row levels of \code{x}.}
}
\value{
A matrix containing the coordinates of the cluster centers.
}
\description{
A helper function for computing the coordinates of cluster centers (typically used in correspondence analysis).
}

