% Generated by roxygen2 (4.0.2): do not edit by hand
\name{complete_pvpick}
\alias{complete_pvpick}
\title{Complete the Output of \code{pvpick}}
\usage{
complete_pvpick(clusters, labels)
}
\arguments{
\item{clusters}{A clustering by a call to \code{pvpick}.}

\item{labels}{A character vector containing the exhaustive set of levels.}
}
\value{
A list with the singleton clusters inserted at the end (so that the set of clusters is exhaustive).
}
\description{
A helper function to add the missing singleton clusters in the output of \code{pvpick} (from the package \pkg{pvclust}).
}

