% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Ctxt_Dut.csv}
\alias{Ctxt_Dut.csv}
\title{Context Words for seventeen Dutch Words}
\format{A table with 17 rows and 1404 columns.}
\description{
The frequency table of seventeen Dutch synonyms of \emph{beginnen} ("to begin") and their context words (from the Dutch Parallel Corpus).
}
\examples{
Ctxt_Dut <- read.csv2(system.file("extdata", "Ctxt_Dut.csv", package = "svs"), row.names = 1)
sca.Ctxt_Dut <- fast_sca(data.matrix(Ctxt_Dut))
sca.Ctxt_Dut
lsa.Ctxt_Dut <- fast_lsa(data.matrix(Ctxt_Dut))
lsa.Ctxt_Dut
}

