% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Ctxt_Eng.csv}
\alias{Ctxt_Eng.csv}
\title{Context Words for seventeen Dutch Words Translated from French}
\format{A table with 17 rows and 609 columns.}
\description{
The frequency table of seventeen Dutch synonyms of \emph{beginnen} ("to begin") and their context words in texts translated from English
  (from the Dutch Parallel Corpus).
}
\examples{
Ctxt_Eng <- read.csv2(system.file("extdata", "Ctxt_Eng.csv", package = "svs"), row.names = 1)
sca.Ctxt_Eng <- fast_sca(data.matrix(Ctxt_Eng))
sca.Ctxt_Eng
lsa.Ctxt_Eng <- fast_lsa(data.matrix(Ctxt_Eng))
lsa.Ctxt_Eng
}

