% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Ctxt_Fra.csv}
\alias{Ctxt_Fra.csv}
\title{Context Words for seventeen Dutch Words Translated from French}
\format{A table with 17 rows and 612 columns.}
\description{
The frequency table of seventeen Dutch synonyms of \emph{beginnen} ("to begin") and their context words in texts translated from French
  (from the Dutch Parallel Corpus).
}
\examples{
Ctxt_Fra <- read.csv2(system.file("extdata", "Ctxt_Fra.csv", package = "svs"), row.names = 1)
sca.Ctxt_Fra <- fast_sca(data.matrix(Ctxt_Fra))
sca.Ctxt_Fra
lsa.Ctxt_Fra <- fast_lsa(data.matrix(Ctxt_Fra))
lsa.Ctxt_Fra
}

