% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svs.r
\docType{data}
\name{Ctxt_Fra.txt}
\alias{Ctxt_Fra.txt}
\title{Context Words for seventeen Dutch Words Translated from French}
\format{A table with 17 rows and 612 columns.}
\description{
The frequency table of seventeen Dutch synonyms of \emph{beginnen} ("to begin") and their context words in texts translated from French
  (from the Dutch Parallel Corpus).
}
\examples{
Ctxt_Fra <- read.table(system.file("extdata", "Ctxt_Fra.txt", package = "svs"),
   header = TRUE, sep = "\t", quote = "\"", encoding = "UTF-8")
sca.Ctxt_Fra <- fast_sca(data.matrix(Ctxt_Fra))
sca.Ctxt_Fra
lsa.Ctxt_Fra <- fast_lsa(data.matrix(Ctxt_Fra))
lsa.Ctxt_Fra
}
