% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svs.r
\name{dat2ind}
\alias{dat2ind}
\title{Transform a Data Frame into an Indicator Matrix}
\usage{
dat2ind(dat, add_names = TRUE)
}
\arguments{
\item{dat}{A data frame.}

\item{add_names}{Logical specifying whether to add rownames to the resulting indicator matrix.}
}
\value{
An indicator matrix.
}
\description{
A helper function for transforming a data frame into an indicator matrix.
}
\seealso{
\code{\link{vec2ind}}.
}
