% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dist_wrt_centers}
\alias{dist_wrt_centers}
\title{Compute Distances with respect to Cluster Centers}
\usage{
dist_wrt_centers(x, clusters, freq = NULL, members_only = TRUE)
}
\arguments{
\item{x}{A numeric matrix (containing coordinates).}

\item{clusters}{A clustering of the row levels of \code{x}: either a list or the output of \code{kmeans}.}

\item{freq}{An optional vector of frequency counts for the row levels of \code{x}.}

\item{members_only}{Logical specifying whether the distances from the cluster centers should only be computed for
the cluster members.}
}
\value{
A list with a matrix of distances for every cluster.
}
\description{
A function for computing (euclidean) distances with respect to specified cluster centers.
}

