% Generated by roxygen2 (4.0.2): do not edit by hand
\name{freq_ca}
\alias{freq_ca}
\title{Compute Level Frequencies (for a Factor)}
\usage{
freq_ca(dat, nfac = FALSE)
}
\arguments{
\item{dat}{A factor or a data frame.}

\item{nfac}{Logical indicating whether the number of factors (i.e. the number of columns in \code{dat}) is a divisor for
the level frequencies.}
}
\value{
A vector containing the frequency counts of every level.
}
\description{
A helper function for computing the frequency of each factor level (typically used in correspondence analysis).
}

