% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{InvT_Eng.csv}
\alias{InvT_Eng.csv}
\title{Seventeen Dutch Target Words and their English Source Words}
\format{A data frame with 782 rows and 2 variables.
\itemize{
  \item{\code{source_Eng} }{The English source word.}
  \item{\code{target_Dut} }{The Dutch target word.}
}}
\description{
The occurrences of seventeen Dutch synonyms of \emph{beginnen} ("to begin") and their English source words (from the Dutch Parallel Corpus).
}
\examples{
InvT_Eng <- read.csv2(system.file("extdata", "InvT_Eng.csv", package = "svs"))
sca.InvT_Eng <- fast_sca(InvT_Eng)
sca.InvT_Eng
lsa.InvT_Eng <- fast_lsa(InvT_Eng)
lsa.InvT_Eng
}

