% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{InvT_Fra.csv}
\alias{InvT_Fra.csv}
\title{Seventeen Dutch Target Words and their French Source Words}
\format{A data frame with 856 rows and 2 variables.
\itemize{
  \item{\code{source_Fra} }{The French source word.}
  \item{\code{target_Dut} }{The Dutch target word.}
}}
\description{
The occurrences of seventeen Dutch synonyms of \emph{beginnen} ("to begin") and their French source words (from the Dutch Parallel Corpus).
}
\examples{
InvT_Fra <- read.csv2(system.file("extdata", "InvT_Fra.csv", package = "svs"))
sca.InvT_Fra <- fast_sca(InvT_Fra)
sca.InvT_Fra
lsa.InvT_Fra <- fast_lsa(InvT_Fra)
lsa.InvT_Fra
}

