% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MI}
\alias{MI}
\alias{mi}
\title{Mutual Information}
\usage{
MI(x, base = 2)

mi(x, base = 2)
}
\arguments{
\item{x}{A table (i.e. an object which can be converted to an array).}

\item{base}{Numeric specification of the base with respect to which logarithms are computed.}
}
\value{
A numeric value containing the mutual information.
}
\description{
A function for computing the mutual information.
}
\examples{
SndT_Fra <- read.table(system.file("extdata", "SndT_Fra.txt", package = "svs"),
   header = TRUE, sep = "\\t", quote = "\\"", encoding = "UTF-8")
tab.SndT_Fra <- table(SndT_Fra)
MI(tab.SndT_Fra)
}
\seealso{
\code{\link{pmi}}.
}

