% Generated by roxygen2 (4.0.2): do not edit by hand
\name{outerec}
\alias{outerec}
\title{Recursive Application of the Outer Product}
\usage{
outerec(...)
}
\arguments{
\item{...}{The specification of two or more arrays (separated by comma's or contained in a list).}
}
\value{
An array with the outer product of all the arrays specified in \code{...}.
}
\description{
A helper function for computing the outer product of two or more arrays.
}

