% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svs.r
\docType{data}
\name{SndT_Eng.txt}
\alias{SndT_Eng.txt}
\title{Seventeen Dutch Source Words and their English Translations}
\format{
A data frame with 1117 rows and 2 variables.
\describe{
  \item{\code{source_Dut}}{The Dutch source word.}
  \item{\code{target_Eng}}{The English target word.}
}
}
\description{
The occurrences of seventeen Dutch synonyms of \emph{beginnen} ("to begin") and their English translations (from the Dutch Parallel Corpus).
}
\examples{
SndT_Eng <- read.table(system.file("extdata", "SndT_Eng.txt", package = "svs"),
   header = TRUE, sep = "\t", quote = "\"", encoding = "UTF-8",
   stringsAsFactors = FALSE)
sca_SndT_Eng <- fast_sca(SndT_Eng)
sca_SndT_Eng
lsa_SndT_Eng <- fast_lsa(SndT_Eng)
lsa_SndT_Eng
}
