% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tab2dat}
\alias{tab2dat}
\title{Transform a Table into a Data Frame}
\usage{
tab2dat(tab)
}
\arguments{
\item{tab}{A table (i.e. an object which can be converted to an array).}
}
\value{
A data frame.
}
\description{
A helper function for transforming a table into a data frame.
}

