% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svs.r
\name{tab2ind}
\alias{tab2ind}
\title{Transform a Table into an Indicator Matrix}
\usage{
tab2ind(tab, sep = "_", add_names = TRUE)
}
\arguments{
\item{tab}{A table or (sparse) matrix.}

\item{sep}{Character specifying the separator string for joining the levels.}

\item{add_names}{Logical specifying whether to add dimnames to the resulting indicator matrix.}
}
\value{
An indicator matrix.
}
\description{
A helper function for transforming a table into an indicator matrix.
}
