% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vec2ind}
\alias{vec2ind}
\title{Transform a Vector into an Indicator Matrix}
\usage{
vec2ind(x, add_names = TRUE)
}
\arguments{
\item{x}{A vector (which will be converted to a factor).}

\item{add_names}{Logical specifying whether to add dimnames to the resulting indicator matrix.}
}
\value{
An indicator matrix.
}
\description{
A helper function for transforming a vector into an indicator matrix.
}
\details{
This is essentially the function \code{class.ind} from the package \pkg{MASS}.
}

