% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svs.r
\docType{data}
\name{Ctxt_Dut.txt}
\alias{Ctxt_Dut.txt}
\title{Context Words for seventeen Dutch Words}
\format{
A table with 17 rows and 1404 columns.
}
\description{
The frequency table of seventeen Dutch synonyms of \emph{beginnen} ("to begin") and their context words (from the Dutch Parallel Corpus).
}
\examples{
Ctxt_Dut <- read.table(system.file("extdata", "Ctxt_Dut.txt", package = "svs"),
   header = TRUE, sep = "\t", quote = "\"", encoding = "UTF-8")
sca.Ctxt_Dut <- fast_sca(data.matrix(Ctxt_Dut))
sca.Ctxt_Dut
lsa.Ctxt_Dut <- fast_lsa(data.matrix(Ctxt_Dut))
lsa.Ctxt_Dut
}
