% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svs.r
\name{tab2dat}
\alias{tab2dat}
\title{Transform a Table into a Data Frame}
\usage{
tab2dat(tab)
}
\arguments{
\item{tab}{A table or (sparse) matrix.}
}
\value{
A data frame.
}
\description{
A helper function for transforming a table into a data frame.
}
