% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tab2ind}
\alias{tab2ind}
\title{Transform a Table into an Indicator Matrix}
\usage{
tab2ind(tab, sep = "_", add_names = TRUE)
}
\arguments{
\item{tab}{A table (i.e. an object which can be converted to an array).}

\item{sep}{Character specifying the separator string for joining the levels.}

\item{add_names}{Logical specifying whether to add dimnames to the resulting indicator matrix.}
}
\value{
An indicator matrix.
}
\description{
A helper function for transforming a table into an indicator matrix.
}

