% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svs.r
\name{vec2ind}
\alias{vec2ind}
\title{Transform a Vector into an Indicator Matrix}
\usage{
vec2ind(x, add_names = TRUE)
}
\arguments{
\item{x}{A vector (which will internally be converted to a factor).}

\item{add_names}{Logical specifying whether to add dimnames to the resulting indicator matrix.}
}
\value{
An indicator matrix.
}
\description{
A helper function for transforming a vector into an indicator matrix.
}
\details{
As of version 2.0.x of the \code{svs} package, this is essentially a wrapper for 
\code{t(fac2sparse())} from the \pkg{Matrix} package.
}
\seealso{
\code{\link{dat2ind}}.
}
